[fix]momberのIcon部分をcommonに置き換え
