--- src/components/pages/common/Members/MemberDetail/MemberDetail.types.ts
@@ -1,5 +1,0 @@
-import { User } from '@/types/application/user.types';
-
-export type MemberDetailProps = {
-  user: User;
-};
