--- src/components/pages/common/Members/MemberDetail/index.ts
@@ -1,1 +1,0 @@
-export { default as MemberDetail } from './MemberDetail';
