--- src/components/pages/common/Members/MemberIcon/MemberIcon.module.scss
@@ -1,120 +1,0 @@
-@use '@/styles/mixin' as *;
-
-.iconContainer {
-  position: relative;
-  width: 16rem;
-  height: 16rem;
-  cursor: pointer;
-
-  @include mq(md) {
-    width: 14rem;
-    height: 14rem;
-  }
-
-  > .iconContainerInner {
-    > img {
-      transition: all 0.3s ease;
-    }
-  }
-
-  &:hover {
-    > .icon {
-      > span:first-child {
-        width: 45%;
-        height: 45%;
-      }
-      > span:last-child {
-        width: 85%;
-        height: 85%;
-      }
-    }
-    > .iconContainerInner {
-      > img {
-        transform: scale(1.1);
-        opacity: 0.8;
-      }
-    }
-  }
-}
-
-.iconContainerInner {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-  overflow: hidden;
-}
-
-.icon {
-  position: relative;
-  display: block;
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-  overflow: hidden;
-  background-color: var(--b3);
-
-  > span:first-child {
-    position: absolute;
-    width: 40%;
-    height: 40%;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    border-radius: 50%;
-    background-color: var(--w1);
-    transition: opacity 0.3s ease, width 0.3s ease, height 0.3s ease;
-  }
-
-  > span:last-child {
-    position: absolute;
-    width: 80%;
-    height: 80%;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, 20%);
-    border-radius: 50%;
-    background-color: var(--w1);
-    transition: opacity 0.3s ease, width 0.3s ease, height 0.3s ease;
-  }
-}
-
-.backgroundOne {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 17rem;
-  height: 17rem;
-  border-radius: 50%;
-  background-color: var(--w1);
-  z-index: -3;
-  transform: translate(-50%, -50%);
-  opacity: 0.3;
-
-  @include mq(md) {
-    width: 15rem;
-    height: 15rem;
-  }
-}
-
-.backgroundTwo {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  width: 18rem;
-  height: 18rem;
-  border-radius: 50%;
-  background-color: var(--w1);
-  z-index: -3;
-  transform: translate(-50%, -50%);
-  opacity: 0.3;
-
-  @include mq(md) {
-    width: 16rem;
-    height: 16rem;
-  }
-}
-
-.iconImage {
-  object-fit: cover;
-}
