--- src/components/pages/common/Members/MemberIcon/MemberIcon.types.ts
@@ -1,5 +1,0 @@
-import { User } from '@/types/application/user.types';
-
-export type MemberIconProps = {
-  user: User;
-};
