--- src/components/pages/common/Members/MemberIcon/index.ts
@@ -1,1 +1,0 @@
-export { default as MemberIcon } from './MemberIcon';
